move breakdown_timedelta to a utils file
