+++ backoffice/views.py
@@ -32,6 +32,7 @@ from drivers.models import (
 )
 from hardware.models import Vehicle, Box, BoxAction, Telemetry
 from users.models import User
+from backoffice.utils import breakdown_timedelta
 
 from .decorators import require_backoffice_access
 from .forms import DriverProfileApprovalForm, DriverProfileReviewForm, CloseBookingForm
@@ -557,47 +558,6 @@ def vehicle_details(request, vehicle_id):
         "bookings": [last_booking, current_booking, next_booking],
     }
     return render(request, "backoffice/vehicles/details.html", context)
-
-
-def breakdown_timedelta(timedelta):
-    seconds_in_an_hour = 60 * 60
-    seconds_in_a_day = 24 * seconds_in_an_hour
-    seconds_in_a_month = 30 * seconds_in_a_day
-    seconds_in_a_year = 365 * seconds_in_a_day
-
-    all_seconds = timedelta.days * seconds_in_a_day + timedelta.seconds
-    years, seconds_remaining = divmod(all_seconds, seconds_in_a_year)
-    months, seconds_remaining = divmod(seconds_remaining, seconds_in_a_month)
-    days, seconds_remaining = divmod(seconds_remaining, seconds_in_a_day)
-    hours, seconds_remaining = divmod(seconds_remaining, seconds_in_an_hour)
-    minutes, seconds_remaining = divmod(seconds_remaining, 60)
-    seconds = int(seconds_remaining)
-    if years > 0:
-        years_str = f"{years} years, "
-    else:
-        years_str = ""
-    if months > 0:
-        months_str = f"{months} months, "
-    else:
-        months_str = ""
-    if days > 0:
-        days_str = f"{days} days, "
-    else:
-        days_str = ""
-    if hours > 0:
-        hours_str = f"{hours} hours, "
-    else:
-        hours_str = ""
-    if minutes > 0:
-        minutes_str = f"{minutes} minutes, "
-    else:
-        minutes_str = ""
-    if seconds > 0:
-        seconds_str = f"and {seconds} seconds"
-    else:
-        seconds_str = ""
-
-    return f"{years_str}{months_str}{days_str}{hours_str}{minutes_str}{seconds_str}"
 
 
 @require_backoffice_access
