+++ hardware/models.py
@@ -4,6 +4,7 @@ import uuid
 from datetime import timedelta
 from django.core.serializers.json import DjangoJSONEncoder
 from django.db import models
+from backoffice.utils import breakdown_timedelta
 
 from users.models import User
 
@@ -256,46 +257,6 @@ class Telemetry(models.Model):
             unit_index += 1
         return f"{free_bytes:.2f} {units[unit_index]}"
 
-    def uptime_to_str(self):
-        if self.box_uptime_s is None:
-            return None
-        seconds_in_an_hour = 60 * 60
-        seconds_in_a_day = 24 * seconds_in_an_hour
-        seconds_in_a_month = 30 * seconds_in_a_day
-        seconds_in_a_year = 365 * seconds_in_a_day
-        years, seconds_remaining = divmod(self.box_uptime_s, seconds_in_a_year)
-        months, seconds_remaining = divmod(seconds_remaining, seconds_in_a_month)
-        days, seconds_remaining = divmod(seconds_remaining, seconds_in_a_day)
-        hours, seconds_remaining = divmod(seconds_remaining, seconds_in_an_hour)
-        minutes, seconds_remaining = divmod(seconds_remaining, 60)
-        seconds = int(seconds_remaining)
-        if years > 0:
-            years_str = f"{years} years, "
-        else:
-            years_str = ""
-        if months > 0:
-            months_str = f"{months} months, "
-        else:
-            months_str = ""
-        if days > 0:
-            days_str = f"{days} days, "
-        else:
-            days_str = ""
-        if hours > 0:
-            hours_str = f"{hours} hours, "
-        else:
-            hours_str = ""
-        if minutes > 0:
-            minutes_str = f"{minutes} minutes, "
-        else:
-            minutes_str = ""
-        if seconds > 0:
-            seconds_str = f"and {seconds} seconds"
-        else:
-            seconds_str = ""
-
-        return f"{years_str}{months_str}{days_str}{hours_str}{minutes_str}{seconds_str}"
-
     def __str__(self):
         return f"""telemetry for box {self.box.id} at {self.created_at}: 
     {self.odometer_miles} miles
@@ -303,7 +264,7 @@ class Telemetry(models.Model):
     aux battery @ {self.aux_battery_voltage}v
     {self.soc_percent}% charged
     ibutton_id: {self.ibutton_id}
-    uptime: {self.uptime_to_str}
+    uptime: {breakdown_timedelta(self.box_uptime_s)}
     free heap: {self.free_heap_bytes_to_str}"""
 
     class Meta:
